# swagger/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Online Payment Platform API",
    "contact": {
      "email": "nbondarczuk@yahoo.com"
    },
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
    "/health": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "health"
        ],
        "summary": "Check server status.",
        "operationId": "HealthHandler",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/payment": {
      "post": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "payment"
        ],
        "summary": "Creates a new payment.",
        "operationId": "CreatePaymentHandler",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/payment/{id}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "payment"
        ],
        "summary": "Reads one one payment by id.",
        "operationId": "ReadOnePaymentHandler",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the payment",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/refund/{id}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "refund"
        ],
        "summary": "Reads one refund by id.",
        "operationId": "ReadOneRefundHandler",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the refund",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "refund"
        ],
        "summary": "Creates a new refund using reference to the original payment.",
        "operationId": "CreateRefundHandler",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}